v0.2.0b preRelease
Misc updates/improvements …

+++ Classes/Controller/BaseController.php
@@ -37,7 +37,7 @@ use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
  * Class BaseController
  * @package S3b0\EcomConfigCodeGenerator\Controller
  */
-class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController {
+class BaseController extends \Ecom\EcomToolbox\Controller\ActionController {
 
 	/**
 	 * @var \S3b0\EcomConfigCodeGenerator\Domain\Model\Content
@@ -101,6 +101,22 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 	protected $frontendUserRepository;
 
 	/**
+	 * regionRepository
+	 *
+	 * @var \Ecom\EcomToolbox\Domain\Repository\RegionRepository
+	 * @inject
+	 */
+	protected $regionRepository;
+
+	/**
+	 * stateRepository
+	 *
+	 * @var \Ecom\EcomToolbox\Domain\Repository\StateRepository
+	 * @inject
+	 */
+	protected $stateRepository;
+
+	/**
 	 * @var \S3b0\EcomConfigCodeGenerator\Domain\Model\Configuration
 	 */
 	protected $configuration = NULL;
@@ -139,7 +155,7 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 		// Frontend-Session
 		$this->feSession->setStorageKey(Setup::getSessionStorageKey($this->contentObject));
 		// On reset destroy config session data
-		if ( $this->request->getControllerActionName() === 'reset' ) {
+		if ( $this->request->getControllerName() === 'Generator' && $this->request->getControllerActionName() === 'reset' ) {
 			$this->feSession->delete('config');
 			$this->redirectToUri($this->uriBuilder->build());
 		}
@@ -198,6 +214,7 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 			 */
 			if ( $partGroup->isLocked() ) {
 				if ( $partGroup->hasDefaultPart() ) {
+					$partGroup->setSelectable(FALSE);
 					\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::addPartToConfiguration(
 						$this,
 						$partGroup->getDefaultPart() instanceof \TYPO3\CMS\Extbase\Persistence\Generic\LazyLoadingProxy ? $partGroup->getDefaultPart()->_loadRealInstance() : $partGroup->getDefaultPart(),
@@ -205,7 +222,7 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 					);
 					$locked++;
 					$cycle++;
-				} else {
+				} elseif ( !array_key_exists($partGroup->getUid(), $configuration) ) {
 					$partGroups->detach($partGroup);
 				}
 				continue;
@@ -248,33 +265,29 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 				// First of all check dependencies and unset parts, if not valid anymore
 				foreach ( $configuration[$partGroup->getUid()] as $partUid ) {
 					/** @var \S3b0\EcomConfigCodeGenerator\Domain\Model\Part $part */
-					$part = $this->partRepository->findByUid($partUid);
-					if ( self::checkForPartDependencies($this, $part, $configuration) ) {
-						\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::addPartToConfiguration($this, $part, $configuration);
-					} else {
-						\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::removePartFromConfiguration($this, $part, $configuration);
+					if ( $part = $this->partRepository->findByUid($partUid) ) {
+						if ( self::checkForPartDependencies($this, $part, $configuration) ) {
+							\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::addPartToConfiguration($this, $part, $configuration);
+						} else {
+							\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::removePartFromConfiguration($this, $part, $configuration);
+						}
 					}
 				}
-				$originallyAvailablePartsAmount = $partGroup->getParts()->count();
-				$this->initializeParts(
-					$partGroup->getParts(),
-					$configuration
-				);
-				/**
-				 * Reset part if availability has been affected by dependencies
-				 */
-				if ( $partGroup->getParts()->count() !== $originallyAvailablePartsAmount && $partGroup->getParts()->count() === 1 ) {
-					$partGroup->setSelectable(FALSE);
-					\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::removePartGroupFromConfiguration($this, $partGroup, $configuration);
-					\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::addPartToConfiguration($this, $partGroup->getParts()->toArray()[0], $configuration);
-				}
-				$partGroup->setUnlocked($partGroup->getParts()->count() > 1);
-				if ( !$partGroup->isUidInParts($configuration[$partGroup->getUid()]) ) {
-					$current = $partGroup;
-				}
-			} else {
-				$current = $partGroup;
 			}
+			$originallyAvailablePartsAmount = $partGroup->getParts()->count();
+			$this->initializeParts(
+				$partGroup->getParts(),
+				$configuration
+			);
+			/**
+			 * Reset part if availability has been affected by dependencies
+			 */
+			if ( $partGroup->getParts()->count() !== $originallyAvailablePartsAmount && $partGroup->getParts()->count() === 1 ) {
+				$partGroup->setSelectable(FALSE);
+				\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::removePartGroupFromConfiguration($this, $partGroup, $configuration);
+				\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::addPartToConfiguration($this, $partGroup->getParts()->toArray()[0], $configuration);
+			}
+			$partGroup->setUnlocked($partGroup->getParts()->count() > 1);
 			$partGroup->setNext($previous);
 			/** @var \S3b0\EcomConfigCodeGenerator\Domain\Model\PartGroup $previous */
 			$previous = $partGroup;
@@ -283,16 +296,14 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 		$cycle = 0;
 		foreach ( $partGroups as $partGroup ) {
 			$partGroup->setStepIndicator($partGroup->isVisibleInNavigation() ? ++$cycle : 0);
+			if ( !array_key_exists($partGroup->getUid(), $configuration) && !is_array($configuration[$partGroup->getUid()]) && !$current instanceof \S3b0\EcomConfigCodeGenerator\Domain\Model\PartGroup ) {
+				$current = $partGroup;
+			}
 		}
 		$this->feSession->store('config', $configuration);
 
 		// Get progress state update (ratio of active to visible packages) => float from 0 to 1 (*100 = %)
 		$progress = ( sizeof($configuration) - $locked ) / ( $partGroups->count() - $locked );
-
-		// Initially use first package
-		if ( $progress < 1 && !$current instanceof \S3b0\EcomConfigCodeGenerator\Domain\Model\PartGroup ) {
-			$current = $partGroup;
-		}
 
 		return $partGroups;
 	}
@@ -402,6 +413,7 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 	 */
 	protected function getConfigurationCode(array $configuration) {
 		$summaryTableRows = [ ];
+		$summaryTableMailRows = [ ];
 		$code = [ ];
 		$blankCode = [ ];
 		/** @var \S3b0\EcomConfigCodeGenerator\Domain\Model\PartGroup $partGroup */
@@ -429,7 +441,11 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 					<td>{$partGroup->getStepIndicator()}</td>
 					<td>{$partGroup->getTitle()}</td>
 					<td>" . implode(', ', $partList) . "</td>
-					<td>" . ($partGroup->isSelectable() ? "<a data-package=\"{$partGroup->getUid()}\" class=\"configurator-part-group-select\"><i class=\"fa fa-edit\"></i></a>" : "") . "</td>
+					<td>" . ($partGroup->isSelectable() ? "<a data-ccgpg=\"{$partGroup->getUid()}\" class=\"configurator-part-group-select\"><i class=\"fa fa-edit\"></i></a>" : "") . "</td>
+				");
+				$summaryTableMailRows[] = ("
+					<td>{$partGroup->getTitle()}</td>
+					<td>" . implode(', ', $partList) . "</td>
 				");
 			}
 		}
@@ -439,17 +455,16 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 		return [
 			'code' => $this->contentObject->getCcgConfiguration()->getPrefix() . implode('', $code) . $this->contentObject->getCcgConfiguration()->getSuffix(),
 			'blankCode' => $this->contentObject->getCcgConfiguration()->getPrefix() . implode('', $blankCode) . $this->contentObject->getCcgConfiguration()->getSuffix(),
-			'summaryTable' => '<table><tr>' . implode('</tr><tr>', $summaryTableRows) . '</tr></table>'
+			'summaryTable' => $this->sanitize_output('<table><tr>' . implode('</tr><tr>', $summaryTableRows) . '</tr></table>'),
+			'summaryTableMail' => $this->sanitize_output('<table><tr>' . implode('</tr><tr>', $summaryTableMailRows) . '</tr></table>')
 		];
 	}
 
 	/**
-	 * @param array $configuration
-	 * @throws \TYPO3\CMS\Extbase\Persistence\Exception\IllegalObjectTypeException
+	 * @param \S3b0\EcomConfigCodeGenerator\Domain\Model\Log $log
+	 * @param array                                          $configuration
 	 */
-	protected function addConfigurationLog(array $configuration) {
-		/** @var \S3b0\EcomConfigCodeGenerator\Domain\Model\Log $log */
-		$log = $this->objectManager->get(\S3b0\EcomConfigCodeGenerator\Domain\Model\Log::class);
+	protected function addConfigurationToLog(\S3b0\EcomConfigCodeGenerator\Domain\Model\Log &$log, array $configuration) {
 		$ipLength = !MathUtility::canBeInterpretedAsInteger($this->settings['log']['ipLength']) || $this->settings['log']['ipLength'] > 4 ? 4 : $this->settings['log']['ipLength'];
 		$code = [ ];
 
@@ -472,21 +487,38 @@ class BaseController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController 
 		}
 		ksort($code);
 
-		$log->setConfiguration($this->contentObject->getCcgConfiguration()->getPrefix() . implode('', $code) . $this->contentObject->getCcgConfiguration()->getSuffix())
-			->maskIpAddress($ipLength);
+		$log->setSessionId($GLOBALS['TSFE']->fe_user->id)
+			->setConfiguration($this->contentObject->getCcgConfiguration()->getPrefix() . implode('', $code) . $this->contentObject->getCcgConfiguration()->getSuffix())
+			->maskIpAddress($ipLength)
+			->setPid(0);
 		if ( $GLOBALS['TSFE']->loginUser ) {
+			/** @var \TYPO3\CMS\Extbase\Domain\Model\FrontendUser $feUser */
+			$feUser = $this->frontendUserRepository->findByUid($GLOBALS['TSFE']->fe_user->user['uid']);
 			/** @todo Add pricing */
-			$log->setFeUser($this->frontendUserRepository->findByUid($GLOBALS['TSFE']->fe_user->user['uid']));
+			$log->setFeUser($feUser);
 		}
+	}
 
-		// Write to DB and persist to obtain a uid for current log
-		$this->logRepository->add($log);
-		/** @var \TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager $persistenceManager */
-		$persistenceManager = $this->objectManager->get(\TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager::class);
-		$persistenceManager->add($log);
-		$persistenceManager->persistAll();
+	/**
+	 * @param string $templateName template name (UpperCamelCase)
+	 * @param array $variables variables to be passed to the Fluid view
+	 *
+	 * @return string
+	 */
+	protected function getStandAloneTemplate($templateName, array $variables = []) {
+		/** @var \TYPO3\CMS\Fluid\View\StandaloneView $view */
+		$view = $this->objectManager->get(\TYPO3\CMS\Fluid\View\StandaloneView::class);
 
-		\S3b0\EcomConfigCodeGenerator\Session\ManageConfiguration::resetConfiguration($this);
+		$extbaseFrameworkConfiguration = $this->configurationManager->getConfiguration(\TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface::CONFIGURATION_TYPE_FRAMEWORK);
+		$templateRootPath = \TYPO3\CMS\Core\Utility\GeneralUtility::getFileAbsFileName($extbaseFrameworkConfiguration['view']['templateRootPath'] ?: end($extbaseFrameworkConfiguration['view']['templateRootPaths']));
+		$partialRootPath = \TYPO3\CMS\Core\Utility\GeneralUtility::getFileAbsFileName($extbaseFrameworkConfiguration['view']['partialRootPath'] ?: end($extbaseFrameworkConfiguration['view']['partialRootPaths']));
+		$templatePathAndFilename = "{$templateRootPath}{$templateName}.html";
+		$view->setTemplatePathAndFilename($templatePathAndFilename);
+		$view->setPartialRootPaths([$partialRootPath]);
+		$view->assignMultiple($variables);
+		$view->setFormat('html');
+
+		return $this->sanitize_output($view->render());
 	}
 
 	/**
